# contracts/openapi/tab-v1.yaml
openapi: 3.0.3
info:
  title: UTab API v1
  version: 0.1.1
  description: >
    Contract-first API used by tournament operations. This is the initial
    machine-readable contract for Phase 0.
servers:
  - url: https://tab.example.com
security:
  - sessionAuth: []
  - bearerAuth: []
...
paths:
  /api/v1/auth/service-token-revocations:
    get:
      tags: [Auth]
      summary: List revoked service-account tokens
      operationId: listServiceTokenRevocations
      parameters:
        - name: active
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeServiceTokenRevocationList"
    post:
      tags: [Auth]
      summary: Revoke service-account token by jti
      operationId: revokeServiceToken
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTokenRevocationCreateRequest"
      responses:
        "200":
          description: Already revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeServiceTokenRevocation"
        "201":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeServiceTokenRevocation"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Required for service-account authenticated POST/PATCH/DELETE requests.
...
  schemas:
    EnvelopeServiceTokenRevocation:
      type: object
      required: [data, errors]
      properties:
        data:
          $ref: "#/components/schemas/ServiceTokenRevocation"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
    EnvelopeServiceTokenRevocationList:
      type: object
      required: [data, errors]
      properties:
        data:
          type: object
          required: [items, limit]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ServiceTokenRevocation"
            limit:
              type: integer
              minimum: 1
              maximum: 200
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
...
    ServiceTokenRevocation:
      type: object
      properties:
        _id:
          type: string
        jti:
          type: string
        reason:
          type: string
        revokedBy:
          type: string
        revokedAt:
          type: string
          format: date-time
        expireAt:
          type: string
          format: date-time
    ServiceTokenRevocationCreateRequest:
      type: object
      required: [jti]
      properties:
        jti:
          type: string
          minLength: 1
          maxLength: 128
        reason:
          type: string
          minLength: 1
          maxLength: 500
        expiresAt:
          type: string
          format: date-time
...
    ErrorItem:
      type: object
      required: [name, message]
      properties:
        name:
          type: string
        message:
          type: string